start by using actual mempage implementation

+++ internal/backend/pager/pager.go
@@ -4,7 +4,7 @@ package pager
  Pager is the interface with which popsql will access the database files
 */
 type Pager interface {
-	ReadPage(pageNum uint16) ([]byte, error)
-	WritePage(pageNum uint16, data []byte) error
+	ReadPage(pageNum uint32) ([]byte, error)
+	WritePage(pageNum uint32, data []byte) error
 	PageSize() uint16
 }
